bot-test/pyproject.toml
[tool.poetry]
name = "bot_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-decouple = "^3.8"
transformers = "^4.46.1"
assemblyai = "^0.34.0"
torch = "^2.5.1"
sentencepiece = "^0.2.0"
python-telegram-bot = {extras = ["webhooks"], version = "^21.7"}
psycopg2-binary = "^2.9.10"
slack-sdk = "^3.33.3"
pyjokes = "^0.8.3"
slack-bolt = "^1.21.2"
jira = "^3.8.0"
requests = "^2.32.3"
botbuilder-integration-aiohttp = "^4.16.2"
openai = "^1.54.4"
gtts = "^2.5.4"
flask = "^3.1.0"
nest-asyncio = "^1.6.0"
docx = "^0.2.4"
spacy = "^3.8.2"
qdrant-client = "^1.12.1"
sentence-transformers = "^3.3.1"
pypdf2 = "^3.0.1"
pdfplumber = "^0.11.4"
python-docx = "^1.1.2"
apscheduler = "^3.11.0"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-magic = "^0.4.27"
python-magic-bin = "^0.4.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
